+++ DistFiles/releaseNotes.md
@@ -1,3 +1,16 @@
+## Version 3.7
+Enabled suppression of repeated short filename warnings.
+Added language chooser to make it easy to select the desired language in a form that supports sharing research and archiving.
+Corrected diplay of person details.
+Prevented waveform scrolling to the beginning after listening to a source segment.
+Fixed display problem in project metadata when no project language has been chosen.
+Fixed a bug when setting the font in the transcription annotation column.
+Added check when saving metadata and EAF files to ensure that they do not end up empty or otherwise corrupt.
+Security patch.
+
+## Version 3.6.2
+Prevented NullReferenceException when clicking away during playback of generated oral annotations.
+
 ## Version 3.6.1
 Fixed problem that prevented creation of a new project.